OVR for flashlight

+++ Assets/Scripts/Hand.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Hand : MonoBehaviour {
+    public Vector3 holdPosition = new Vector3(0, -0.025f, 0.03f);
+    public Vector3 holdRotation = new Vector3(0, 180, 0);
     public OVRInput.Controller controller;
 
     private float indexState = 0;
@@ -10,6 +12,7 @@ public class Hand : MonoBehaviour {
 
     private bool grabbing = false;
     private ChessPiece chessPiece = null;
+    private FlashLight flashLight = null;
 
     public void Update() {
         indexState = OVRInput.Get(OVRInput.Axis1D.PrimaryIndexTrigger, controller);
@@ -27,6 +30,11 @@ public class Hand : MonoBehaviour {
             if (!grabbing && indexState > 0.9f) {
                 Grab(other.gameObject);
             }
+        } else if (other.CompareTag("FlashLight")) {
+             Debug.Log(grabbing + " " + indexState);
+             if (!grabbing && indexState > 0.9f) {
+                 GrabFlashLight(other.gameObject);
+             }
         }
     }
 
@@ -36,9 +44,33 @@ public class Hand : MonoBehaviour {
         chessPiece.pickUp(gameObject);
     }
 
+    private void GrabFlashLight(GameObject flashlight) {
+        grabbing = true;
+        // flashLight = flashlight.GetComponent<FlashLight>();
+        flashLight = flashlight;
+        flashLight.transform.parent = transform;
+        gun.transform.localPosition = holdPosition;
+        gun.transform.localEulerAngles = holdRotation;
+        gun.GetComponent<Rigidbody>().useGravity = false;
+        gun.GetComponent<Rigidbody>().isKinematic = true;
+    }
+
     private void Release() {
         grabbing = false;
-        chessPiece.drop();
-        chessPiece = null;
+        if (chessPiece )
+        {
+            chessPiece.drop();
+            chessPiece = null;
+        } 
+        else 
+        {
+            flashLight.transform.parent = null;
+            Rigidbody rigidbody = flashLight.GetComponent<Rigidbody>();
+            rigidbody.useGravity = true;
+            rigidbody.isKinematic = false;
+            rigidbody.velocity = OVRInput.GetLocalControllerVelocity(controller);
+            grabbing = false;
+            flashLight = null;
+        }
     }
 }
